+++ src/main/java/roomescape/controller/AdminReservationController.java
@@ -1,0 +1,50 @@
+package roomescape.controller;
+
+import java.net.URI;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+import roomescape.dto.request.ReservationRequest;
+import roomescape.dto.response.ReservationResponse;
+import roomescape.service.ReservationCommandService;
+import roomescape.service.ReservationQueryService;
+
+@RestController
+@RequestMapping("/admin/reservations")
+@RequiredArgsConstructor
+public class AdminReservationController {
+
+    private final ReservationCommandService reservationCommandService;
+    private final ReservationQueryService reservationQueryService;
+
+    @GetMapping
+    public ResponseEntity<List<ReservationResponse>> getAllReservations() {
+        return ResponseEntity.ok(reservationQueryService.getAllReservations());
+    }
+
+    @PostMapping
+    public ResponseEntity<ReservationResponse> createReservation(@RequestBody ReservationRequest request) {
+        ReservationResponse reservationResponse = reservationCommandService.create(request.name(), request.date(),
+                request.timeId(), request.themeId());
+
+        URI location = ServletUriComponentsBuilder.fromCurrentRequest()
+                .build()
+                .toUri();
+
+        return ResponseEntity.created(location).body(reservationResponse);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteReservation(@PathVariable("id") Long id) {
+        reservationCommandService.delete(id);
+        return ResponseEntity.noContent().build();
+    }
+}
