+++ src/main/java/roomescape/controller/AdminReservationTimeController.java
@@ -1,0 +1,49 @@
+package roomescape.controller;
+
+import java.net.URI;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+import roomescape.dto.request.ReservationTimeRequest;
+import roomescape.dto.response.ReservationTimeResponse;
+import roomescape.service.ReservationTimeCommandService;
+import roomescape.service.ReservationTimeQueryService;
+
+@RestController
+@RequestMapping("/admin/times")
+@RequiredArgsConstructor
+public class AdminReservationTimeController {
+
+    private final ReservationTimeCommandService reservationTimeCommandService;
+    private final ReservationTimeQueryService reservationTimeQueryService;
+
+    @GetMapping
+    public ResponseEntity<List<ReservationTimeResponse>> getAllTimes() {
+        return ResponseEntity.ok(reservationTimeQueryService.findAllReservationTimes());
+    }
+
+    @PostMapping
+    public ResponseEntity<ReservationTimeResponse> createReservationTime(@RequestBody ReservationTimeRequest request) {
+        ReservationTimeResponse reservationTimeResponse = reservationTimeCommandService.create(request.startAt());
+
+        URI location = ServletUriComponentsBuilder.fromCurrentRequest()
+                .build()
+                .toUri();
+
+        return ResponseEntity.created(location).body(reservationTimeResponse);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteReservationTime(@PathVariable("id") Long id) {
+        reservationTimeCommandService.delete(id);
+        return ResponseEntity.noContent().build();
+    }
+}
