+++ src/main/java/roomescape/controller/AdminThemeController.java
@@ -1,0 +1,50 @@
+package roomescape.controller;
+
+import java.net.URI;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+import roomescape.dto.request.ThemeRequest;
+import roomescape.dto.response.ThemeResponse;
+import roomescape.service.ThemeCommandService;
+import roomescape.service.ThemeQueryService;
+
+@RestController
+@RequestMapping("/admin/themes")
+@RequiredArgsConstructor
+public class AdminThemeController {
+
+    private final ThemeCommandService themeCommandService;
+    private final ThemeQueryService themeQueryService;
+
+    @GetMapping
+    public ResponseEntity<List<ThemeResponse>> getAllThemes() {
+        return ResponseEntity.ok(themeQueryService.findAllThemes());
+    }
+
+    @PostMapping
+    public ResponseEntity<ThemeResponse> createTheme(@RequestBody ThemeRequest request) {
+        ThemeResponse themeResponse = themeCommandService.create(request.name(), request.thumbnailUrl(),
+                request.description());
+
+        URI location = ServletUriComponentsBuilder.fromCurrentRequest()
+                .build()
+                .toUri();
+
+        return ResponseEntity.created(location).body(themeResponse);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteTheme(@PathVariable("id") Long id) {
+        themeCommandService.delete(id);
+        return ResponseEntity.noContent().build();
+    }
+}
